fix: sign_up issue and admob fixed

+++ lib/bet_slip/widgets/bet_slip_card.dart
@@ -378,7 +378,8 @@ class _BetSlipCardState extends State<BetSlipCard> {
                                             await UserRepository()
                                                 .rewardForVideoAd(
                                               uid: currentUserId,
-                                              rewardValue: rewardItem.amount,
+                                              rewardValue:
+                                                  rewardItem.amount.toInt(),
                                             );
                                           };
                                           final ads = RewardAd(
